detection_logic.py
import time
import random
import os
from PIL import Image
import PIL
import json # To format results/errors

def run_photo_manipulation_detection(filepath):
    """
    Placeholder detection logic (no changes needed inside here).
    Returns a dictionary with results or raises an Exception on failure.
    """
    print(f"Processing image file: {filepath}")
    try:
        img = Image.open(filepath)
        img.verify()
        img.close() # Close file handle after verify
        img = Image.open(filepath) # Re-open
        print(f"Image Info - Format: {img.format}, Size: {img.size}, Mode: {img.mode}")
        img.close() # Close file handle after getting info

    except (IOError, SyntaxError, PIL.UnidentifiedImageError) as e:
        print(f"Error opening or verifying image {filepath}: {e}")
        raise ValueError(f"Invalid or corrupted image file: {e}") from e
    except Exception as e:
        print(f"Unexpected error processing image {filepath}: {e}")
        raise

    # Simulate processing time
    processing_time = random.uniform(2, 8)
    time.sleep(processing_time)

    # Simulate detection result
    possible_types = ['AI-Generated (GAN)', 'Likely Manipulated', 'Likely Real']
    detected_type = random.choice(possible_types)
    is_manipulated = detected_type != 'Likely Real'
    confidence = random.uniform(0.6, 0.99) if is_manipulated else random.uniform(0.01, 0.4)

    result = {
        'is_manipulated': is_manipulated,
        'confidence': round(confidence, 4),
        'suspected_type': detected_type,
        'message': f"Image analysis complete."
    }
    print(f"Detection result for {filepath}: {result}")
    return result

def format_error_result(e, original_filename=""):
     """ Formats an exception into a dictionary for storing in DB """
     return json.dumps({
          'status': 'error',
          'original_filename': original_filename,
          'error_type': type(e).__name__,
          'error_message': str(e)
     })

def format_success_result(detection_result, original_filename=""):
    """ Formats the successful detection result for storing in DB """
    return json.dumps({
        'original_filename': original_filename,
        'detection': detection_result,
        'status': 'success'
    })